Iterate and refine send and active transaction behavior
Send is now functional and shows up in active transactions.
This involved:
- reducing the exposure of the seed
- consistently using Ints for blockheight everywhere to match zcash
- adding the use of spending keys
- adding account initialization on startup
- using accounts but defaulting to account 0
- internalizing birthday so we no longer need a reference outside of the library
- enabling cancellation during send
- cleanup active transaction manager

+++ src/main/java/cash/z/wallet/sdk/dao/BlockDao.kt
@@ -1,7 +1,7 @@
 package cash.z.wallet.sdk.dao
 
 import androidx.room.*
-import cash.z.wallet.sdk.vo.Block
+import cash.z.wallet.sdk.entity.Block
 
 @Dao
 interface BlockDao {
@@ -24,7 +24,7 @@ interface BlockDao {
     fun deleteAll()
 
     @Query("SELECT MAX(height) FROM blocks")
-    fun lastScannedHeight(): Long
+    fun lastScannedHeight(): Int
 
     @Query("UPDATE blocks SET time=:time WHERE height = :height")
     fun updateTime(height: Int, time: Int)
