+++ src/main/java/cash/z/wallet/sdk/ext/WalletService.kt
@@ -2,8 +2,8 @@ package cash.z.wallet.sdk.ext
 
 import cash.z.wallet.sdk.rpc.Service
 
-inline fun Long.toBlockHeight(): Service.BlockID = Service.BlockID.newBuilder().setHeight(this).build()
-inline fun LongRange.toBlockRange(): Service.BlockRange =
+inline fun Int.toBlockHeight(): Service.BlockID = Service.BlockID.newBuilder().setHeight(this.toLong()).build()
+inline fun IntRange.toBlockRange(): Service.BlockRange =
     Service.BlockRange.newBuilder()
         .setStart(this.first.toBlockHeight())
         .setEnd(this.last.toBlockHeight())
